+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.34 ] - 2025/01/26
+
+  * Added card configuration option `deviceDefaultId` for specifying a Device ID to use for all SpotifyPlus service calls that require a deviceId argument.  This allows you to hide the Devices section, and utilize the specific device for all service requests that require a deviceId.
+  * Added card configuration option `playerVolumeMaxValue` for specifying the maximum volume value allowed to be set via the card user-interface.  This value does not apply if adjusting the volume via services or other media player UI's.
+
 ###### [ 1.0.33 ] - 2025/01/23
 
   * This release requires the SpotifyPlus Integration v1.0.86+ release; please make sure you update the SpotifyPlus integration prior to updating this SpotifyPlus Card release.